Refactor folder name
