+++ Builders/Bill Ball/Sumo 4/Files for 3D printing/blade.stl
