+++ Builders/Graham_Home/code/code.py
@@ -1,4 +1,4 @@
-from .base_bot import SumoBotBase
+from base_bot import SumoBotBase
 
 
 class GrahamSumoBot(SumoBotBase):
@@ -46,21 +46,22 @@ class GrahamSumoBot(SumoBotBase):
                 and not (self.right_edge_detected() or self.left_edge_detected())
             ):
                 self.drive(right_speed=1, left_speed=1)
+            self.stop()
 
         # Check for enemy to right
         elif self.enemy_in_range_right():
             # Turn to the right for 0.5 seconds
-            self.drive(left_speed=-0.5, right_speed=0.5, duration=0.5)
+            self.drive(left_speed=-0.5, right_speed=0.5, duration=1)
 
         # Check for enemy to left
         elif self.enemy_in_range_left():
             # Turn to the left for 0.5 seconds
-            self.drive(left_speed=0.5, right_speed=-0.5, duration=0.5)
+            self.drive(left_speed=0.5, right_speed=-0.5, duration=1)
 
         # Spin and look for enemy
         else:
             # Turn to the left for 0.5 seconds
-            self.drive(left_speed=0.5, right_speed=-0.5, duration=0.25)
+            self.drive(left_speed=0.5, right_speed=-0.5, duration=1)
 
 
 if __name__ == "__main__":
